refactor: create base fragment for fragment favorite
To reduce duplication and improve maintenance
