--- feature/favorite/src/main/res/layout/fragment_favorite_movies.xml
@@ -1,56 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.coordinatorlayout.widget.CoordinatorLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:id="@+id/coordinator_layout"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent">
-
-  <!-- Main list -->
-  <androidx.swiperefreshlayout.widget.SwipeRefreshLayout
-    android:id="@+id/swipe_refresh"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <androidx.recyclerview.widget.RecyclerView
-      android:id="@+id/rv_fav_movies"
-      android:layout_width="match_parent"
-      android:layout_height="wrap_content"
-      android:clipToPadding="false"
-      android:paddingTop="10dp"
-      android:paddingBottom="?attr/actionBarSize"
-      app:layout_behavior="@string/appbar_scrolling_view_behavior"
-      tools:listitem="@layout/item_result" />
-
-  </androidx.swiperefreshlayout.widget.SwipeRefreshLayout>
-
-  <!-- Illustration for error state (e.g., network error) -->
-  <include
-    android:id="@+id/illustration_error"
-    layout="@layout/illustration_error"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_gravity="center"
-    android:layout_marginBottom="@dimen/margin_to_centered"
-    android:visibility="gone"
-    tools:visibility="visible" />
-
-  <!-- Illustration for "no data found" state -->
-  <include
-    android:id="@+id/illustration_no_data_view"
-    layout="@layout/illustration_no_data"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_gravity="center"
-    android:layout_marginBottom="@dimen/margin_to_centered"
-    android:visibility="gone"
-    tools:visibility="invisible" />
-
-  <ProgressBar
-    android:id="@+id/progress_bar"
-    style="@style/MyProgressBar"
-    android:layout_gravity="center"
-    android:layout_marginBottom="@dimen/margin_to_centered"
-    tools:visibility="visible" />
-
-</androidx.coordinatorlayout.widget.CoordinatorLayout>
